requirements.txt
pandas
numpy
scikit-learn
joblib
cryptography
paho-mqtt
psutil
